requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GitHub API
PyGithub==2.1.1
requests==2.31.0

# RAG & Embeddings
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
chromadb==0.4.22
faiss-cpu==1.12.0

# LLM Clients
openai==1.6.1
anthropic==0.8.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Web Interface (Week 4+)
streamlit==1.29.0
plotly==5.18.0

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality (for development)
black==23.12.1
ruff==0.1.9
mypy==1.7.1